backend/fabrique/requirements.txt
django==4.1
djangorestframework==3.14
gunicorn==20.1.0
python-dotenv==0.20
celery==5.2.7
requests==2.28.1
psycopg2-binary==2.9.5
drf-yasg==1.21.4
